support windows for the container.scala, communicate between process with Name Pipe.
